# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechMarket Algeria API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  discountPercent?: number | null;
  categoryId: number;
  categoryName: string;
  brandId: number;
  brandName: string;
  imageUrl: string;
  images: string[];
  rating: number;
  reviewCount: number;
  inStock: boolean;
  stockQuantity: number;
  featured: boolean;
  isNew: boolean;
  tags: string[];
  createdAt: string;
}

export interface Specification {
  group: string;
  key: string;
  value: string;
}

export type ProductDetail = Product & {
  specifications: Specification[];
  relatedProducts: Product[];
};

export interface ProductsListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Subcategory {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  productCount: number;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  description: string;
  imageUrl: string;
  icon: string;
  productCount: number;
  subcategories: Subcategory[];
}

export interface Brand {
  id: number;
  name: string;
  slug: string;
  logoUrl: string;
  description: string;
  country: string;
  productCount: number;
  featured: boolean;
}

export interface Promotion {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  discount: string;
  imageUrl: string;
  validUntil: string;
  categoryId?: number | null;
  active: boolean;
}

export interface CartItem {
  productId: number;
  productName: string;
  brandName: string;
  imageUrl: string;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface WishlistInput {
  productId: number;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  cash_on_delivery: "cash_on_delivery",
  cib: "cib",
  edahabia: "edahabia",
} as const;

export interface OrderInput {
  customerName: string;
  phone: string;
  email?: string;
  wilaya: string;
  commune: string;
  address: string;
  paymentMethod: OrderInputPaymentMethod;
  items: CartItemInput[];
  notes?: string | null;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  customerName: string;
  wilaya: string;
  total: number;
  createdAt: string;
}

export interface Review {
  id: number;
  productId: number;
  authorName: string;
  rating: number;
  title: string;
  body: string;
  verified: boolean;
  createdAt: string;
}

export type ListProductsParams = {
  category?: string;
  brand?: string;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  featured?: boolean;
  onSale?: boolean;
  inStock?: boolean;
  sortBy?: ListProductsSortBy;
  page?: number;
  limit?: number;
};

export type ListProductsSortBy =
  (typeof ListProductsSortBy)[keyof typeof ListProductsSortBy];

export const ListProductsSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  rating: "rating",
  popular: "popular",
} as const;

export type ListReviewsParams = {
  productId: number;
};