# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechMarket Algeria API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: categories
    description: Category operations
  - name: brands
    description: Brand operations
  - name: cart
    description: Cart operations
  - name: wishlist
    description: Wishlist operations
  - name: orders
    description: Order operations
  - name: promotions
    description: Promotion operations
  - name: reviews
    description: Review operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: brand
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: onSale
          in: query
          schema:
            type: boolean
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, newest, rating, popular]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsListResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/new-arrivals:
    get:
      operationId: getNewArrivals
      tags: [products]
      summary: Get new arrivals
      responses:
        "200":
          description: New arrivals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/best-sellers:
    get:
      operationId: getBestSellers
      tags: [products]
      summary: Get best sellers
      responses:
        "200":
          description: Best sellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /brands:
    get:
      operationId: listBrands
      tags: [brands]
      summary: List all brands
      responses:
        "200":
          description: Brands list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"

  /promotions:
    get:
      operationId: listPromotions
      tags: [promotions]
      summary: List active promotions
      responses:
        "200":
          description: Promotions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared

  /cart/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "200":
          description: Added to wishlist

  /wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove from wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from wishlist

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        discountPercent:
          type: integer
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        brandId:
          type: integer
        brandName:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        featured:
          type: boolean
        isNew:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - nameAr
        - slug
        - description
        - price
        - categoryId
        - categoryName
        - brandId
        - brandName
        - imageUrl
        - images
        - rating
        - reviewCount
        - inStock
        - stockQuantity
        - featured
        - isNew
        - tags
        - createdAt

    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            specifications:
              type: array
              items:
                $ref: "#/components/schemas/Specification"
            relatedProducts:
              type: array
              items:
                $ref: "#/components/schemas/Product"
          required:
            - specifications
            - relatedProducts

    Specification:
      type: object
      properties:
        group:
          type: string
        key:
          type: string
        value:
          type: string
      required:
        - group
        - key
        - value

    ProductsListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        icon:
          type: string
        productCount:
          type: integer
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
      required:
        - id
        - name
        - nameAr
        - slug
        - description
        - imageUrl
        - icon
        - productCount
        - subcategories

    Subcategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - nameAr
        - slug
        - productCount

    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
        description:
          type: string
        country:
          type: string
        productCount:
          type: integer
        featured:
          type: boolean
      required:
        - id
        - name
        - slug
        - logoUrl
        - description
        - country
        - productCount
        - featured

    Promotion:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        discount:
          type: string
        imageUrl:
          type: string
        validUntil:
          type: string
          format: date-time
        categoryId:
          type: integer
          nullable: true
        active:
          type: boolean
      required:
        - id
        - title
        - titleAr
        - description
        - discount
        - imageUrl
        - validUntil
        - active

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - deliveryFee
        - total
        - itemCount

    CartItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        brandName:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
      required:
        - productId
        - productName
        - brandName
        - imageUrl
        - price
        - quantity
        - subtotal

    CartItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity

    CartItemUpdate:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity

    WishlistInput:
      type: object
      properties:
        productId:
          type: integer
      required:
        - productId

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
        wilaya:
          type: string
        commune:
          type: string
        address:
          type: string
        paymentMethod:
          type: string
          enum: [cash_on_delivery, cib, edahabia]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
        notes:
          type: string
          nullable: true
      required:
        - customerName
        - phone
        - wilaya
        - commune
        - address
        - paymentMethod
        - items

    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        customerName:
          type: string
        wilaya:
          type: string
        total:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - status
        - customerName
        - wilaya
        - total
        - createdAt

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        authorName:
          type: string
        rating:
          type: integer
        title:
          type: string
        body:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - authorName
        - rating
        - title
        - body
        - verified
        - createdAt