# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechMarket Algeria API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  customerName: string;
  wilaya: string;
  total: number;
  createdAt: Date;
}