# lib/api-zod/src/generated/types/specification.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechMarket Algeria API
 * OpenAPI spec version: 0.1.0
 */

export interface Specification {
  group: string;
  key: string;
  value: string;
}